Home

Trello.jl

Trello.TrelloCred

Trello.create_board

Trello.create_card

Trello.create_list

Trello.delete_board

Trello.get_boards

Trello.get_cards

Trello.get_labels

Trello.get_lists

Trello.request

Trello.TrelloCred — Type.

TrelloCred(;api_key, api_token)

A credentials and identity object for Trello. If the apikey or apitoken are not provided, they will be read from the TRELLO_API_KEY and TRELLO_API_TOKEN environment variables respectively.

Follow the instructions in the Trello Documentation to aquire your credentials.

source

Trello.create_board — Method.

create_board(cred::TrelloCred, name; desc="", id_organization="")

Create a new board on the cred account. Board is created completely empty (no lists, nor labels). If id_organization is provided then it will be created within that organization.

source

Trello.create_card — Method.

create_card(cred::TrelloCred, list_id, name; desc="", label_ids::Vector=[])

Create a card, on the given list with the name, description and labels as specified.

See Trello docs:

https://developers.trello.com/reference#cards-2

https://api.trello.com/1/cards?idList=idList&keepFromSource=all

source

Trello.create_list — Function.

create_list(credit::TrelloCred, board_id, name, pos="bottom")

Create a list within the given board, with the given name, at position given by pos

source

Trello.delete_board — Method.

delete_board(cred::TrelloCred, id)

Deletes the board.

source

Trello.get_boards — Method.

get_boards(cred::TrelloCred)

List all Trello boards on your account. (On the account assoicated with the credentials.)

source

Trello.get_cards — Method.

get_cards(cred::TrelloCred, list_id)

Returns all the cards on a given list.

source

Trello.get_labels — Method.

get_labels(cred::TrelloCred, board_id)

Get all the labels within a given Trello board.

source

Trello.get_lists — Method.

get_lists(cred::TrelloCred, board_id)

Get all the lists within a given Trello board.

source

Trello.request — Method.

request(method, cred, path; query_kwargs...)

Send a request to Trello and get back the reponse.

method is the HTTP method to use, e.g "GET", "POST", "PUT" or "DELETE" the get_request, post_request, put_request and delete_request are convenience wrapppers of request with the relevant method.

cred: a TrelloCred object for the account making the request.

path the endpoint being targetted, e.g. "/1/boards/5d721712fa52696736da60cf/lists". Often you will need to interpolate a board or list id or similar into this. (like 5d721712fa52696736da60cf)

query_kwargs: the query to be used with the request, e.g idList="5d72175351f2b73b0f872e03", name="My new Card", desc="The card description", ...

When invoking request it is important to look at the Trello API docs to workout what arguements are allowed, and how they are spelled etc. https://developers.trello.com/reference

Returns the parsed response. Generally as a NamedTuple, or a Vector of NamedTuples

Throws errors if there are HTTP errors.

source